refactor repair pass
